Lookup pldm bootside without targeting for pnor i/o path
Hostboot cannot perform attribute lookups in the path that performs
PNOR file i/o request, otherwise we can end up in a deadlock situation.
To avoid this, we will use a variable scoped to the PLDM namespace
which will be checked during file i/o requests. It will be set lazily,
ie. if its not set we will look it up on the fly and store the value
we find for future file io accesses.

Change-Id: Iad078f9ee5660c0db09bf28d337c9ecbd4e80c69
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118897
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>