RAS Review: p10_hcd_*.xml error xmls
This commit goes through all of the p10_hcd_*.xml files and checks
that all of the ffdc fields are being populated, and that the callouts
will work. A common problem was that many errors were referencing this
PROC_CHIP_IN_ERROR target which was not defined anywhere. The solution
is to instead use the multicase target this is available and lookup
the core targets associated with it. Aside from this problem there
were a few instances where FFDC fields that were not being populated
and also this commit moves away from using the HCD_ASSERT macro which
has limitation to the number of FFDC tags we can use and also might cause
headaches in the future when we try to add some static analysis to ensure
all ffdc fields are being filled out.

Change-Id: If5dae30805ed436f143aae7070644cfdf8111c1c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118931
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: YUE DU <daviddu@us.ibm.com>
Reviewed-by: Gregory S Still <stillgs@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118998
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>