Fix multicast target generation for special wakeup call
Code was incorrectly creating the multicast target that gets
passed into the p10_core_special_wakeup HWP.  This results in
some unexpected scom errors showing up inside the procedure.

The change above exposed an issue related to mpipl which
requires us to skip the wakeup calls that the PM code
issues in the mpipl path.

Change-Id: I2282bc3b55c2fe05b39dbc0097e03140ba91bdf8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118971
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>