Refactor PLDM Sensor/Effecter ID storage
BACKGROUND

Previously, each target that was related to a PLDM sensor or effecter
had two attributes called PLDM_SENSOR_INFO and
PLDM_EFFECTER_INFO. These attributes stored information about the
sensor or effecter that was "attached" to the target, like the
sensor/effecter ID, state set ID, and function callback ID. For
example, the OCC state sensor would be attached to the OCC
target. This data is stored in attributes because it is generated by
Hostboot at IPL time, but is also needed by HBRT to service PLDM
requests, and attributes are a natural mechanism for communicating
data from IPL code to runtime code. (Anything that lives in Hostboot's
normal memory space, like global variables, is not accessible to
HBRT. Attributes are stored in a special memory region that is
preserved in the transition.)

This approach has a few notable downsides:

1. Each target can have at most one sensor and one effecter.
2. Each sensor/effecter must be associated with a target, even if
   it is not related to any particular target.
3. Adding a new sensor/effecter requires adding an attribute to a
   target rather than just writing code, which is annoying.

THIS COMMIT

To solve these constraints, this commit introduces a new system-level
volatile attribute called PLDM_STATE_QUERY_RECORDS. This attribute is
used as backing storage for an array of records that describe sensors
and effecters. This way, a target can have any number of
sensors/effecters associated with it; each sensor/effecter doesn't
have to be associated with a target; and adding a new sensor/effecter
doesn't require changing targeting. The old PLDM_SENSOR_INFO and
PLDM_EFFECTER_INFO have been removed and the code refactored to use
this new attribute to store sensor/effecter information.

Change-Id: I11de6373c053b373bc8c78f2b3ce6d1ca3613ce4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119215
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>