Update hb-memdump.sh to work on FSP-B
The hb-memdump.sh script uses multiple ecmds to dump hostboot's memory.
These commands rely on being run from the Primary FSP and knowing the
position number of the boot proc to be successful.

The boot proc is in position 0 for the ecmds when FSP-A is the Primary FSP
or when running on an eBMC system.

This commit adds a check so that position 1 is used for the boot proc
for the ecmds when FSP-B is the Primary FSP.

Change-Id: I3d25d034cfeb4743b75e2e688b2035e38f44c3c7
CQ:SW528549
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119239
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>