Clear I2C Atomic Locks During Hostboot IPL
This commit takes the support to clear the I2C Atomic Locks from
the HBRT code and uses it during the Hostboot part of the IPL.  Now,
when hostboot normally does the I2C resets for all I2C engines during
the IPL it will also clear any Atomic Locks on the I2C engine.

This commit also adds FFDC for when an I2C register access hits an
XSCOM_STATUS_ERR in order to see who might be holding an atomic lock
on the I2C engine.

In addition, this code cleans up a few minor things from the previous
commit that introduced clearing the I2C Atomic Locks into the HBRT
code.

Change-Id: I52f6d1d0506d173f9dda9149688a5290ae54bf5d
CQ:SW528362
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119505
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>