Add delay between OCMB reset deassert and first i2c operation
We have encountered situations where the first i2c operation to
Explorer seems to lock up the chip.  The theory is that sometimes
the Microchip firmware takes longer to get to a stable point to
handle i2c operations so a delay will avoid this problem.  The
delay is set to 10 seconds by default.

Change-Id: I85e06d12327f21264a7c3a029a56e08567f19a6a
CQ: SW527725
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119557
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>