Fix setGardType and associated tests
This commit addresses several issues with setGardType and
its associated testcases.

1. setGardType was comparing ud callout type with gard type which are
not comparable and resulted in the function being a noop.
2. The testcases that could have caught this issue weren't verifying the
gard type was actually set but instead using just bit fields to detect
that a gard was set.
3. The bit field for GARD_FOUND type was overlapping with TARGET_MATCH
and DECONFIG_FOUND so that as long as the requested target was found in
the callouts the test case would always think it saw a gard record when
there wasn't one.
4. The if statements responsible for all this checking were not properly
formed such that all possible scenarios would be verified. This masked
the above issues from being detected by the testcases.

Change-Id: I55823882c5acf7ac6fd20454eadf19b97a84fd16
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119891
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>