Change callout type constants to an enum
The callout types for an error log have been a set of constants for some
time. However, when switched to an enum type it allows the compiler to
help find improper usages of the callout types. One example is with
setGardType where it was comparing callout type to GARD error type, an
incompatible comparison.

It will also allow switch statements lacking a default attribute to
catch missing callout types in it and throw a compilation error.

Change-Id: I9db44c85eab1c7516c351e4525ff6ac7a3a195cd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119967
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>