Update the comment in tpmPresence to be more precise
In tpmPresence there was a comment that suggested that we should never
gard a TPM. However, there was also a comment in
crosscheck_sp_presence_target that suggested otherwise. This commit
changes the wording in tpmPresence to be more consistent with what is
being done in crosscheck_sp_presence_target.

Change-Id: Ib0ecbb08208e2317f669d7bd4aa5f7b97c7982f3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120102
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>