MPIPL: Increase the memory for MPIPL earlier in the IPL process
At the beginning of the MPIPL process, the memory is not getting
updated properly and it is only updated much later in the
IPL process.  This is causing an issue with Everest which needs
much more memory upfront therefore moved the code that updates
the memory to the start of the IPL process.

Change-Id: I0ce4daf37895ad9da20912f72cca11c652b33694
RTC: 256823
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120118
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>