Allow returns from updateProcessorSbeSeeproms
There can be a situation where updateProcessorSbeSeeproms gets called
but the SBE image currently running matches the one we customized, so
no reboot occurs and the function returns. This can happen when we
detect a scratch reg mismatch in a case where a target is garded that
HB is not aware of. Currently the code that calls
updateProcessorSbeSeeproms throws errors when it returns. This commit
reduces the severity of those errors to informational and commits
the errors so they don't fail the boot.

Change-Id: Ic7bf63482eb2cbe7b4499e1229750778c1e3363d
CQ: SW530746
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120424
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120544
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>