Guard and Deconfigure Unprovisioned TPMs
This commit looks for the specific return code that indicates that hostboot
is interacting with an unprovisioned TPM.  If the TPM Required policy is
enabled then the TPM will be guarded and deconfigured.

This will help the logic of the FSP code that determines which FSP to boot
from, and thereby determines which TPMs are being used.

Please note that Hostboot will only possibly see this fail when the
node-to-node communication procedure is run on a multinode system.

Change-Id: I543f26d51bf177a277f7994267cf8c5279c13977
CQ:SW528740
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120539
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>