Write SRC or EID in Shutdown Path
The logic in doShutdown/shutdown path always assumed that the incoming
shutdown status is an SRC, but doShutdown can be called with an EID as
well, which leads to bogus HB SRCs reported in the TI area. This commit
adds logic to differentiate between SCR and EID and call the appropriate
functions accordingly.

Note that some calls to termWriteSRC have been left intact, since in
some situations we want the final TI code to be overwritten (like if a
kernel crash happens during a shutdown) with the more severe SRC.

Change-Id: I3de86aa6d3cfbb21c78f192fd421a4bb7cb61be4
CQ: SW531141
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120619
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>