PAU Frequency depends only on service processor
The value of ATTR_FREQ_PAU_MHZ is set by code that runs on
both the Service Processor (FSP/BMC) and inside Hostboot.
That means that the value in the scratch reg is not affected
by what HB pushes down in our sync.  At a steady-state both
sides will be running the same level of code, but there are
windows when we could be running different logic that produces
different values.  To handle that case, we will always let
the SP's value win by forcing HB's copy to match what the SP
set into the SBE.

Change-Id: I7d3d90cc902936328a99b39d3416387959540d16
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119148
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>