update FFDC error handling in SbeFifo::readResponse
This commit corrects some of the error handling when we receive
a response from the SBE that contains FFDC. If the FFDC contains
a scom PIB error and addFruCallouts() is called on the error log,
we will now skip adding automatic callouts after all of the FFDC
packets have been processed. This is so we don't add unnecessary
callouts to the error log which might interfere with diagnosing
defects related to these errors.

This commit also links the plids of any HWP error logs generated
by FAPI_SET_SBE_ERROR and fapi2::rcToErrl() to the inital
SBEIO_FIFO_RESPONSE_ERROR.

Change-Id: Ief04a9f700247058273946ceeb0a5f33886dc044
CQ: SW522090
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120888
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>