Change HWAS algorithm to allow processors to be connected by a single SMPGROUP
Previously, the HWAS algorithm required all the SMPGROUP links between
processors to be functional, and would deconfigure a processor that
had even a single SMPGROUP deconfigured. This is not the correct
behavior, and the fabric can work with only a single SMPGROUP
connection between processors.

This commit fixes the issue by changing the HWAS algorithm to permit a
processor to remain functional when there is at least one SMPGROUP
connection between it and any other processor.

Change-Id: I581113d3928c6dc16c73b4cb11c9d9cf2718d517
CQ: SW532222
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121542
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>