Allow SMPGROUP guards to unset BLOCK_SPEC_DECONFIG flag
Previously, when an SMPGROUP guard caused the BLOCK_SPEC_DECONFIG
attribute to be set (via deconfiguring a processor or the like), and
the guard was cleared, BLOCK_SPEC_DECONFIG was not being updated
because of the SMPGROUP targets' HWAS state change subscription
flags. This commit updates the flags to allow them to unset the flag
when appropriate.

Change-Id: Id1d59eff4797038c252d8ef0f59fe3124c36d854
CQ: SW531698
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121129
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>