SBE Image Read-Back Verification
This commit introduces a change to read the newly-customized
SBE image back out of SBE SEEPROM to perform integrity checks
(check that the image binary matches what we customized and
make sure ECC is correct). This is to aid the debug of intermittent
SBE ECC failures on OpenPOWER systems. The read is done page-by-page after
the customized image is written to the SEEPROM. This logic only
applies to non-FSP systems.

Change-Id: I8ab4491ab79aac378b8dafe846b7e10f52cb351a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121623
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>