Use only non-ECO cores for core checks
Found two more places where using only the active (non-ECO) cores
makes more sense.  When looking for a target to match our boot
core we should only include cores we can actually execute on.
When trying to figure out the optimal set of scanrings to include
in the SBE image, we would rather include the active cores if
we have to choose.

Change-Id: I447876d34e31dde4187ef7813180432399d25ffd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122726
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>