BMC BIOS: Get the TPM Required Policy and set ATTR_TPM_REQUIRED to value
Get the TPM Required Policy from the BMC BIOS and set the system's
attribute ATTR_TPM_REQUIRED to the retrieved value if no error occurred.

If an error occurs while retrieving the BIOS value then the system's
attribute ATTR_TPM_REQUIRED is left as is and an error log is created stating
the issue.

A value of 1 is interpreted as the TPM is required for the system to boot
whereas a value of of 0 is interpreted as the TPM is *not* required for
the system to boot.

Change-Id: If16ccac3814f8d9098549f16ee687b8ddd722a85
RTC: 261378
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121480
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>