WOFDATA: Temporary SKIP of PNOR/LID handling for eBMC
Currently op-build does not populate a default wof_binary_filename
(which is used to seed the WOFDATA PNOR partition as a Hash Page Table
enabled section, HPT).  Defining the WOFDATA PNOR partition as HPT-enabled
allows the read only data to be swapped out under memory pressure.

When the wof_binary_filename is EMPTY, the WOFDATA PNOR partition
logic in genPnorImages.pl will skip the HPT building of the WOFDATA
partition.  Subsequently if an attempt is made in getOverrideWofTable
to access the WOFDATA PNOR data, the memcpy will cause an OOM
fault on eBMC platforms.

Temporarily we will skip the getOverrideWofTable call on non-FSP
platforms to circumvent this memory pressure until RTC 250794
(PM: Enable WOFDATA generation for op-build) can be fully implemented.

Change-Id: Ic8c8de5a7a98bc45797d491be6d1b0a80513154b
CQ: SW532696
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121705
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>