Make power_limit_enable and power_limit_in_watts always read current values
Whenever the OCC power limit is enabled or disabled, or the limit
itself is changed, the values are immediately sent to the
OCCs. Whenever Hostboot reads the BIOS attributes for these values, we
need them to read out the same values as the OCCs actually have,
rather than the latched attribute value. This commit bypasses the
attribute latching for these two attributes and always returns the
most recent value whenever they are read.

Change-Id: I3124a1f43e678dbe78561ee8cee01d71634122bf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121732
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>