Fix bus deconfiguration between master-capable procs when FSP-B is primary
The HWAS code deconfigures processors based on the bus connections
between them. Whenever there is a deconfigured bus between the master
processor and another processor, that other processor is deconfigured.

Currently, the HWAS code assumes that the first master-capable
processor is the acting master. This is a faulty assumption on FSP
systems when FSP-B is the primary FSP, which appoints processor 1 as
the boot processor. Currently, Hostboot will misidentify processor 0
as the active master when it is functional, and deconfigure processor
1 when there is a non-functional bus between the two. This will cause
the boot to fail in the minimum hardware check (because the system
can't boot when the boot proc is not functional), and the system will
attempt to fail over to FSP-A.

This commit fixes the issue by checking PROC_MASTER_TYPE before
identifying a processor as the active master.

Change-Id: Ibe99bad1d4014fd43e8aaa05fc9ea9bb2095ab8e
CQ: SW533117
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121760
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>