Inhibit SBE halt if not fully supported by SBE
There is a chicken and egg problem with the new SBE halt flow and how FSP/SBE
handle that flow.  In order for that flow to work properly, code from FSP, SBE,
and Hostboot all need to be installed.  However, when updating firmware, the
SBE code may lag behind for awhile, which leads to incompatibilities.

This change will remove the incompatibility across the labs and manufacturing.
If firmware is imprint signed and SBE supports reporting Hostboot requested
SBE halts to the service processor, halt SBEs in istep 10.1 through istep 10.3.
If firmware is imprint signed and SBE does not properly report that halt status,
inhibit the SBE halt processing in those isteps.
If firmware is production signed, always force the SBE halt processing since it
would be a security threat not to do so.

Further, inhibit the SBE un-halt in MPIPL flow, since it's not halted in MPIPL.

Change-Id: I32dfe088b70c23effe0499b0206e784f7d1c442d
CQ: SW533240
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121874
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>