Improve Trace.pm to avoid endless recursion issue
If we take a hostboot dump while hostboot is running, the trace
buffers can be getting written as we are dumping the memory which
can lead to the trace buffers linked list effectively looping
on itself. If this happens the Trace tool will go into an infinite
recursive loop which will continue consuming memory until either
linux crashes or linux kills the program. To avoid this, keep
track of previous addresses we have walked, and if we try to
read an address we have already looked at, break out of the
recursion.

Change-Id: I3aed857a6b279b7188f32c266bc2daf835b42272
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121887
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>