Correctly process target type from single attr dump cronus request
In P10, the fapi2 target type enumeration changed from a uint32_t
to a bitwise uint64_t. Now each bit in the uint64_t represents a
different target type. The broke the single attribute dump request
format as it operated on uint32_t only. To get around this, cronus
now sends the bit that we need to set in the uint64_t as a uint32_t.
This commit changes the logic on the Hostboot side to take the uint32_t
cronus sends and translate it to the new bitwise uint64_t format for
fapi2 target types.

Change-Id: I6cb52291c9b201acd65934234d7ee783b98c4829
CQ: SW532686
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/121958
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>