Do not read sbe seeprom side from boot proc
The implementation of the SBE halt workaround during
step10 of the IPL triggered a dormant error scenario
path in the Hostboot code.  This path uses the boot
processor's register to determine which seeprom was
used to boot the SBE.  In this case, the boot processor
was on a different seeprom than the 3 secondary procs,
so the answer we get back is incorrect.  This leads us
to boot onto the same side that we updated (good) but
never update the alternate side (bad).

The fix is to make sure we do not use this alternate
code path, but instead always do a scom to the correct
processor.

Change-Id: Icb74fee040563a082a3afccea9eae068e3ff0d41
CQ: SW533544
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122023
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>