SBE_SELECT_EX_POLICY persistency
Make ATTR_SBE_SELECT_EX_POLICY volatile.

Situation which surfaced was when a system had this attribute
set acccording to some previous development driver and this
attribute once set is defined as nonvolatile+writeable, so
once this was set and not re-flashed the value did not get
properly updated to the current default defined, e.g. HB_MAX_FOOTPRINT.

Change-Id: I73ecee4751b50f76dfb133f7ebec84cdeb780419
CQ: SW533565
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122180
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>